test(epub): extract readable fallback family
