+++ src/research/epub.test.ts
@@ -1,7 +1,6 @@
 import { createHash } from 'node:crypto'
 import { strFromU8, strToU8, unzipSync, zipSync } from 'fflate'
 import { XMLParser, XMLValidator } from 'fast-xml-parser'
-import { PDFDocument, rgb } from 'pdf-lib'
 import { describe, expect, it } from 'vitest'
 import { fixtureFile } from '../../tests/fixtures/pdf-fixtures'
 import rawPaper from './papers/semantic-responsive-typesetting.json'
@@ -9,12 +8,10 @@ import {
   buildEpub,
   buildReadableEpub,
   inspectEpub,
-  MAX_EPUB_ASSET_BYTES_PER_BOOK,
-  MAX_READABLE_FALLBACK_EQUATIONS_PER_BOOK,
   profileEpubCss,
-  projectReadableFallbackReconstruction,
   renderPublicationXhtml,
 } from './epub'
+import { staleEquationTranscriptFixture } from './epub-readable-fallback-fixtures'
 import {
   createSourceGeometryScriptTranscript,
   SOURCE_GEOMETRY_SCRIPT_TRANSCRIPT_EVIDENCE,
@@ -38,17 +35,12 @@ import {
   pdfModelConsultationSemanticStateSha256,
   resolvePdfModelFallbacks,
 } from './model-fallback-pipeline'
-import {
-  attachRequiredSourcePageRenders,
-  reconstructPdf,
-  sourcePageRenderBudgetUsage,
-} from './pdf'
+import { reconstructPdf } from './pdf'
 import { reconstructPageAnalyses } from './pdf-layout'
 import { assessPdfCompleteness } from './pdf-quality'
 import { validatedPdfVisualRelationships } from './pdf-visual-validation'
 import { researchPaperSchema } from './schema'
 import { getTargetProfile, resolveTargetProfile } from './targets'
-import { MAX_EPUB_ASSETS_PER_BOOK } from './publication-resource-limits'
 import { createSourcePageCropAsset } from './visual-assets'
 
 const paper = researchPaperSchema.parse(rawPaper)
@@ -387,797 +379,6 @@ function staleNoteAnchorFixture() {
     assets: [],
   } as unknown as PdfReconstruction
   return { notePaper, reconstruction }
-}
-
-async function staleEquationTranscriptFixture() {
-  const captionRun: PdfSourceRun = {
-    page: 1,
-    text: 'Equation 1. Source-backed display.',
-    x: 0.2,
-    y: 0.4,
-    width: 0.42,
-    height: 0.02,
-    rotation: 0,
-    method: 'pdf-text',
-    fontName: 'Body',
-    fontSize: 10,
-    confidence: 1,
-  }
-  const firstRun: PdfSourceRun = {
-    ...captionRun,
-    text: 'q = r',
-    x: 0.28,
-    y: 0.3,
-    width: 0.14,
-    fontSize: 14,
-  }
-  const secondRun: PdfSourceRun = {
-    ...firstRun,
-    text: '+ s',
-    y: 0.325,
-    width: 0.08,
-  }
-  const titleRun: PdfSourceRun = {
-    ...captionRun,
-    text: 'Equation lineage',
-    x: 0.12,
-    y: 0.08,
-    width: 0.36,
-    fontSize: 18,
-  }
-  const authorRun: PdfSourceRun = {
-    ...captionRun,
-    text: 'Ada Researcher',
-    x: 0.12,
-    y: 0.14,
-    width: 0.24,
-    fontSize: 11,
-  }
-  const captionBox = { ...captionRun }
-  const sourceBox = {
-    page: 1,
-    x: 0.27,
-    y: 0.29,
-    width: 0.2,
-    height: 0.065,
-    rotation: 0,
-    method: 'pdf-object' as const,
-  }
-  const page: PdfPageAnalysis = {
-    page: 1,
-    kind: 'born-digital',
-    width: 612,
-    height: 792,
-    rotation: 0,
-    textCharacters: [
-      titleRun,
-      authorRun,
-      captionRun,
-      firstRun,
-      secondRun,
-    ].reduce((total, run) => total + run.text.length, 0),
-    imageCount: 1,
-    objects: [
-      {
-        id: 'equation-source-object',
-        page: 1,
-        kind: 'image',
-        box: sourceBox,
-        confidence: 1,
-        assetId: null,
-      },
-    ],
-    runs: [titleRun, authorRun, firstRun, secondRun, captionRun],
-  }
-  const pixels = new Uint8Array(24 * 12 * 4).fill(255)
-  for (let y = 3; y < 9; y += 1) {
-    for (let x = 4; x < 20; x += 1) {
-      const offset = (y * 24 + x) * 4
-      pixels.set([20, 20, 20, 255], offset)
-    }
-  }
-  const asset = await createSourcePageCropAsset({
-    kind: 'equation',
-    cropBox: sourceBox,
-    sourceObjectIds: ['equation-source-object'],
-    sourceBoxes: [sourceBox],
-    width: 24,
-    height: 12,
-    pixels,
-  })
-  page.objects![0].assetId = asset.id
-  const base = await reconstructPageAnalyses({
-    pages: [page],
-    sourceHash: 'e'.repeat(64),
-    fileName: 'equation-lineage.pdf',
-    byteLength: 1024,
-  })
-  const equationRegion = {
-    id: 'equation-source-region',
-    page: 1,
-    kind: 'equation' as const,
-    column: 'single' as const,
-    text: `${firstRun.text} ${secondRun.text}`,
-    confidence: 1,
-    box: {
-      ...firstRun,
-      width: Math.max(firstRun.width, secondRun.width),
-      height: secondRun.y + secondRun.height - firstRun.y,
-    },
-    lines: [
-      {
-        id: 'equation-source-line-1',
-        text: firstRun.text,
-        fontSize: firstRun.fontSize,
-        box: { ...firstRun },
-        runs: [{ ...firstRun }],
-      },
-      {
-        id: 'equation-source-line-2',
-        text: secondRun.text,
-        fontSize: secondRun.fontSize,
-        box: { ...secondRun },
-        runs: [{ ...secondRun }],
-      },
-    ],
-    nativeObjectIds: [],
-    includedInReadingOrder: true,
-  }
-  const captionRegion = {
-    id: 'equation-caption-region',
-    page: 1,
-    kind: 'caption' as const,
-    column: 'single' as const,
-    text: captionRun.text,
-    confidence: 1,
-    box: { ...captionRun },
-    lines: [
-      {
-        id: 'equation-caption-line',
-        text: captionRun.text,
-        fontSize: captionRun.fontSize,
-        box: { ...captionRun },
-        runs: [{ ...captionRun }],
-      },
-    ],
-    nativeObjectIds: [],
-    includedInReadingOrder: true,
-  }
-  const titleRegion = {
-    id: 'equation-title-region',
-    page: 1,
-    kind: 'spanning' as const,
-    column: 'single' as const,
-    text: titleRun.text,
-    confidence: 1,
-    box: { ...titleRun },
-    lines: [
-      {
-        id: 'equation-title-line',
-        text: titleRun.text,
-        fontSize: titleRun.fontSize,
-        box: { ...titleRun },
-        runs: [{ ...titleRun }],
-      },
-    ],
-    nativeObjectIds: [],
-    includedInReadingOrder: true,
-  }
-  const authorRegion = {
-    id: 'equation-author-region',
-    page: 1,
-    kind: 'body' as const,
-    column: 'single' as const,
-    text: authorRun.text,
-    confidence: 1,
-    box: { ...authorRun },
-    lines: [
-      {
-        id: 'equation-author-line',
-        text: authorRun.text,
-        fontSize: authorRun.fontSize,
-        box: { ...authorRun },
-        runs: [{ ...authorRun }],
-      },
-    ],
-    nativeObjectIds: [],
-    includedInReadingOrder: true,
-  }
-  const equationPaper = structuredClone(paper)
-  equationPaper.title = 'Equation lineage'
-  equationPaper.authors = ['Ada Researcher']
-  equationPaper.nodes = [
-    {
-      id: 'equation-title-node',
-      type: 'heading',
-      level: 1,
-      text: titleRun.text,
-      source: 'synthetic-equation-lineage',
-    },
-    {
-      id: 'equation-author-node',
-      type: 'paragraph',
-      text: authorRun.text,
-      source: 'synthetic-equation-lineage',
-    },
-    {
-      id: 'equation-node',
-      type: 'figure',
-      objectType: 'equation',
-      title: captionRun.text,
-      relationships: {
-        caption: 'equation-caption',
-        assets: [asset.id],
-      },
-      source: 'synthetic-equation-lineage',
-    },
-    {
-      id: 'equation-caption',
-      type: 'caption',
-      text: captionRun.text,
-      source: 'synthetic-equation-lineage',
-    },
-  ]
-  const relationship: PublicationVisualRelationship = {
-    id: 'equation-lineage-relationship',
-    kind: 'equation',
-    label: 'Equation 1',
-    captionRegionId: captionRegion.id,
-    sourceRegionIds: [equationRegion.id],
-    sourceLineIds: equationRegion.lines.map((line) => line.id),
-    sourceObjectIds: ['equation-source-object'],
-    assetIds: [asset.id],
-    status: 'matched',
-    confidence: 1,
-    evidence: ['source-page-crop', 'source-text-alt'],
-    candidates: [],
-    sourceBoxes: [captionBox, sourceBox],
-    sourceText: equationRegion.text,
-    altText: captionRun.text,
-    altTextSource: 'source-text',
-    canonicalNodeId: 'equation-node',
-    captionNodeId: 'equation-caption',
-  }
-  const lineBoundaryDecisions = [
-    {
-      id: 'equation-line-boundary-1',
-      page: 1,
-      regionId: equationRegion.id,
-      fromLineId: equationRegion.lines[0].id,
-      toLineId: equationRegion.lines[1].id,
-      outcome: 'space' as const,
-      evidence: ['synthetic-source-line-order'],
-    },
-  ]
-  const readingOrder = {
-    schemaVersion: '1.0.0' as const,
-    regionIds: [
-      titleRegion.id,
-      authorRegion.id,
-      equationRegion.id,
-      captionRegion.id,
-    ],
-    order: [
-      titleRegion.id,
-      authorRegion.id,
-      equationRegion.id,
-      captionRegion.id,
-    ],
-    edges: [],
-    resolutions: [],
-    acyclic: true,
-    evaluation: {
-      schemaVersion: '1.0.0' as const,
-      algorithm: 'deterministic-geometry-v1' as const,
-      mode: 'deterministic-only' as const,
-      regionCount: 4,
-      acceptedEdgeCount: 0,
-      unresolvedEdgeCount: 0,
-      cycleRate: 0,
-      orderAccuracy: null,
-      provider: null,
-      modelVersion: null,
-      latencyMs: 0 as const,
-      costUsd: 0 as const,
-      reviewRequired: false,
-    },
-  }
-  const candidate = {
-    ...base,
-    paper: equationPaper,
-    pages: [page],
-    regions: [titleRegion, authorRegion, equationRegion, captionRegion],
-    readingOrder,
-    lineBoundaryDecisions,
-    unresolvedCorruptingJoinCount: 0,
-    structurallyConsumedLineBoundaryCount: 0,
-    noteRelationships: [],
-    citationRelationships: [],
-    crossReferenceRelationships: [],
-    visualRelationships: [relationship],
-    assets: [asset],
-    provenance: {
-      'equation-title-node': {
-        confidence: 1,
-        pages: [1],
-        regionIds: [titleRegion.id],
-        boxes: [{ ...titleRun }],
-        links: [],
-      },
-      'equation-author-node': {
-        confidence: 1,
-        pages: [1],
-        regionIds: [authorRegion.id],
-        boxes: [{ ...authorRun }],
-        links: [],
-      },
-      'equation-node': {
-        confidence: 1,
-        pages: [1],
-        regionIds: [equationRegion.id],
-        boxes: relationship.sourceBoxes,
-        links: [],
-      },
-      'equation-caption': {
-        confidence: 1,
-        pages: [1],
-        regionIds: [captionRegion.id],
-        boxes: [captionBox],
-        links: [],
-      },
-    },
-    diagnostics: [],
-  } satisfies PdfReconstruction
-  const assessment = assessPdfCompleteness({
-    pages: candidate.pages,
-    paper: candidate.paper,
-    diagnostics: candidate.diagnostics,
-    readingOrder: candidate.readingOrder,
-    regions: candidate.regions,
-    visualRelationships: candidate.visualRelationships,
-    assets: candidate.assets,
-    citationRelationships: candidate.citationRelationships,
-    noteRelationships: candidate.noteRelationships,
-    policy: candidate.readiness.policy,
-    lineBoundaryDecisions: candidate.lineBoundaryDecisions,
-    sourceSemanticFlowBoundaryDecisions:
-      candidate.sourceSemanticFlowBoundaryDecisions,
-    sourceSemanticFlowBoundaryDecisionCount:
-      candidate.sourceSemanticFlowBoundaryDecisionCount,
-    canonicalHyphenBoundaryDecisions:
-      candidate.canonicalHyphenBoundaryDecisions,
-    canonicalHyphenBoundaryDecisionCount:
-      candidate.canonicalHyphenBoundaryDecisionCount,
-    unresolvedCorruptingJoinCount: candidate.unresolvedCorruptingJoinCount,
-    structurallyConsumedLineBoundaryCount:
-      candidate.structurallyConsumedLineBoundaryCount,
-    provenance: candidate.provenance,
-    sourceSha256: candidate.source.sha256,
-  })
-  const reconstruction = {
-    ...candidate,
-    semanticSignals: assessment.semanticSignals,
-    completeness: assessment.completeness,
-    diagnostics: assessment.diagnostics,
-    readiness: assessment.readiness,
-  } satisfies PdfReconstruction
-  expect(reconstruction.readiness.blockingDiagnosticCodes).toEqual([])
-  return { reconstruction, relationship, equationRegion }
-}
-
-function readableFallbackFigure(index: number) {
-  const suffix = String(index).padStart(3, '0')
-  const bytes = new TextEncoder().encode(`SOURCE-FIGURE-${suffix}`)
-  const sha256 = createHash('sha256').update(bytes).digest('hex')
-  const assetId = `asset-${sha256.slice(0, 24)}`
-  const sourceObjectId = `fallback-figure-object-${suffix}`
-  const captionRegionId = `fallback-figure-caption-region-${suffix}`
-  const nodeId = `fallback-figure-node-${suffix}`
-  const captionNodeId = `fallback-figure-caption-${suffix}`
-  const captionBox = {
-    page: 1,
-    x: 0.1,
-    y: 0.5,
-    width: 0.4,
-    height: 0.02,
-    rotation: 0,
-    method: 'pdf-text' as const,
-  }
-  const sourceBox = {
-    page: 1,
-    x: 0.1,
-    y: 0.55,
-    width: 0.2,
-    height: 0.1,
-    rotation: 0,
-    method: 'pdf-object' as const,
-  }
-  const asset = {
-    id: assetId,
-    href: `assets/${assetId}.png`,
-    mediaType: 'image/png' as const,
-    kind: 'raster' as const,
-    rendition: 'source-preserved' as const,
-    sha256,
-    bytes,
-    width: 20,
-    height: 10,
-    resolutionDpi: null,
-    sourceObjectIds: [sourceObjectId],
-    sourceBoxes: [sourceBox],
-  } satisfies PublicationAsset
-  const captionText = `Figure ${index}. Readable fallback source figure.`
-  const relationship = {
-    id: `fallback-figure-relationship-${suffix}`,
-    kind: 'figure' as const,
-    label: `Figure ${index}`,
-    captionRegionId,
-    sourceRegionIds: [],
-    sourceLineIds: [],
-    sourceObjectIds: [sourceObjectId],
-    assetIds: [asset.id],
-    status: 'matched' as const,
-    confidence: 1,
-    evidence: ['source-preserved-raster'],
-    candidates: [],
-    sourceBoxes: [captionBox, sourceBox],
-    sourceText: '',
-    altText: captionText,
-    altTextSource: 'caption' as const,
-    canonicalNodeId: nodeId,
-    captionNodeId,
-  } satisfies PublicationVisualRelationship
-  return {
-    asset,
-    relationship,
-    nodes: [
-      {
-        id: nodeId,
-        type: 'figure' as const,
-        title: captionText,
-        relationships: {
-          caption: captionNodeId,
-          assets: [asset.id],
-        },
-        source: 'synthetic-readable-fallback-figure',
-      },
-      {
-        id: captionNodeId,
-        type: 'caption' as const,
-        text: captionText,
-        source: 'synthetic-readable-fallback-figure',
-      },
-    ],
-    provenance: {
-      [nodeId]: {
-        confidence: 1,
-        pages: [1],
-        regionIds: [],
-        boxes: relationship.sourceBoxes,
-        links: [],
-      },
-      [captionNodeId]: {
-        confidence: 1,
-        pages: [1],
-        regionIds: [captionRegionId],
-        boxes: [captionBox],
-        links: [],
-      },
-    },
-  }
-}
-
-function withReadableFallbackFigures(
-  reconstruction: PdfReconstruction,
-  figures: ReturnType<typeof readableFallbackFigure>[],
-) {
-  return {
-    ...reconstruction,
-    paper: {
-      ...reconstruction.paper,
-      nodes: [
-        ...reconstruction.paper.nodes,
-        ...figures.flatMap((figure) => figure.nodes),
-      ],
-    },
-    provenance: {
-      ...reconstruction.provenance,
-      ...Object.assign({}, ...figures.map((figure) => figure.provenance)),
-    },
-    visualRelationships: [
-      ...reconstruction.visualRelationships,
-      ...figures.map((figure) => figure.relationship),
-    ],
-    assets: [
-      ...reconstruction.assets,
-      ...figures.map((figure) => figure.asset),
-    ],
-  } satisfies PdfReconstruction
-}
-
-async function readableFallbackEquation(index: number) {
-  const suffix = String(index).padStart(3, '0')
-  const sourceObjectId = `fallback-equation-object-${suffix}`
-  const sourceRegionId = `fallback-equation-region-${suffix}`
-  const sourceLineId = `fallback-equation-line-${suffix}`
-  const captionRegionId = `fallback-equation-caption-region-${suffix}`
-  const nodeId = `fallback-equation-node-${suffix}`
-  const captionNodeId = `fallback-equation-caption-${suffix}`
-  const captionBox = {
-    page: 1,
-    x: 0.2,
-    y: 0.4,
-    width: 0.42,
-    height: 0.02,
-    rotation: 0,
-    method: 'pdf-text' as const,
-  }
-  const lineBox = {
-    page: 1,
-    x: 0.28,
-    y: 0.3,
-    width: 0.14,
-    height: 0.02,
-    rotation: 0,
-    method: 'pdf-text' as const,
-  }
-  const sourceBox = {
-    ...lineBox,
-    x: 0.27,
-    y: 0.29,
-    width: 0.2,
-    height: 0.05,
-    method: 'pdf-object' as const,
-  }
-  const pixels = new Uint8Array(12 * 6 * 4).fill(255)
-  for (let y = 1; y < 5; y += 1) {
-    for (let x = 2; x < 10; x += 1) {
-      pixels.set([20, 20, 20, 255], (y * 12 + x) * 4)
-    }
-  }
-  const asset = await createSourcePageCropAsset({
-    kind: 'equation',
-    cropBox: sourceBox,
-    sourceObjectIds: [sourceObjectId],
-    sourceBoxes: [sourceBox],
-    width: 12,
-    height: 6,
-    pixels,
-  })
-  const sourceText = `q${index} = r`
-  const region = {
-    id: sourceRegionId,
-    page: 1,
-    kind: 'equation' as const,
-    column: 'single' as const,
-    text: sourceText,
-    confidence: 1,
-    box: lineBox,
-    lines: [
-      {
-        id: sourceLineId,
-        text: sourceText,
-        fontSize: 14,
-        box: lineBox,
-        runs: [
-          {
-            ...lineBox,
-            text: sourceText,
-            fontName: 'Synthetic-Math',
-            fontSize: 14,
-            confidence: 1,
-          },
-        ],
-      },
-    ],
-    nativeObjectIds: [],
-    includedInReadingOrder: true,
-  } satisfies PdfReconstruction['regions'][number]
-  const captionText = `Equation ${index}. Readable fallback source equation.`
-  const relationship = {
-    id: `fallback-equation-relationship-${suffix}`,
-    kind: 'equation' as const,
-    label: `Equation ${index}`,
-    captionRegionId,
-    sourceRegionIds: [sourceRegionId],
-    sourceLineIds: [sourceLineId],
-    sourceObjectIds: [sourceObjectId],
-    assetIds: [asset.id],
-    status: 'matched' as const,
-    confidence: 1,
-    evidence: ['source-page-crop', 'source-text-alt'],
-    candidates: [],
-    sourceBoxes: [captionBox, sourceBox],
-    sourceText,
-    altText: captionText,
-    altTextSource: 'source-text' as const,
-    canonicalNodeId: nodeId,
-    captionNodeId,
-  } satisfies PublicationVisualRelationship
-  return {
-    asset,
-    region,
-    relationship,
-    nodes: [
-      {
-        id: nodeId,
-        type: 'figure' as const,
-        objectType: 'equation' as const,
-        title: captionText,
-        relationships: {
-          caption: captionNodeId,
-          assets: [asset.id],
-        },
-        source: 'synthetic-readable-fallback-equation',
-      },
-      {
-        id: captionNodeId,
-        type: 'caption' as const,
-        text: captionText,
-        source: 'synthetic-readable-fallback-equation',
-      },
-    ],
-    provenance: {
-      [nodeId]: {
-        confidence: 1,
-        pages: [1],
-        regionIds: [sourceRegionId],
-        boxes: relationship.sourceBoxes,
-        links: [],
-      },
-      [captionNodeId]: {
-        confidence: 1,
-        pages: [1],
-        regionIds: [captionRegionId],
-        boxes: [captionBox],
-        links: [],
-      },
-    },
-  }
-}
-
-function withReadableFallbackEquations(
-  reconstruction: PdfReconstruction,
-  equations: Awaited<ReturnType<typeof readableFallbackEquation>>[],
-) {
-  return {
-    ...reconstruction,
-    paper: {
-      ...reconstruction.paper,
-      nodes: [
-        ...reconstruction.paper.nodes,
-        ...equations.flatMap((equation) => equation.nodes),
-      ],
-    },
-    provenance: {
-      ...reconstruction.provenance,
-      ...Object.assign({}, ...equations.map((equation) => equation.provenance)),
-    },
-    regions: [
-      ...reconstruction.regions,
-      ...equations.map((equation) => equation.region),
-    ],
-    visualRelationships: [
-      ...reconstruction.visualRelationships,
-      ...equations.map((equation) => equation.relationship),
-    ],
-    assets: [
-      ...reconstruction.assets,
-      ...equations.map((equation) => equation.asset),
-    ],
-  } satisfies PdfReconstruction
-}
-
-function readableFallbackEquationSharingFigureAsset(
-  figure: ReturnType<typeof readableFallbackFigure>,
-) {
-  const sourceRegionId = 'shared-cross-kind-equation-region'
-  const sourceLineId = 'shared-cross-kind-equation-line'
-  const captionRegionId = 'shared-cross-kind-equation-caption-region'
-  const nodeId = 'shared-cross-kind-equation-node'
-  const captionNodeId = 'shared-cross-kind-equation-caption'
-  const sourceBox = figure.asset.sourceBoxes[0]
-  const lineBox = {
-    ...sourceBox,
-    method: 'pdf-text' as const,
-  }
-  const captionBox = {
-    page: 1,
-    x: 0.2,
-    y: 0.4,
-    width: 0.42,
-    height: 0.02,
-    rotation: 0,
-    method: 'pdf-text' as const,
-  }
-  const sourceText = 'q = r'
-  const region = {
-    id: sourceRegionId,
-    page: 1,
-    kind: 'equation' as const,
-    column: 'single' as const,
-    text: sourceText,
-    confidence: 1,
-    box: lineBox,
-    lines: [
-      {
-        id: sourceLineId,
-        text: sourceText,
-        fontSize: 14,
-        box: lineBox,
-        runs: [
-          {
-            ...lineBox,
-            text: sourceText,
-            fontName: 'Synthetic-Math',
-            fontSize: 14,
-            confidence: 1,
-          },
-        ],
-      },
-    ],
-    nativeObjectIds: [],
-    includedInReadingOrder: true,
-  } satisfies PdfReconstruction['regions'][number]
-  const relationship = {
-    id: 'shared-cross-kind-equation-relationship',
-    kind: 'equation' as const,
-    label: 'Equation shared',
-    captionRegionId,
-    sourceRegionIds: [sourceRegionId],
-    sourceLineIds: [sourceLineId],
-    sourceObjectIds: [...figure.asset.sourceObjectIds],
-    assetIds: [figure.asset.id],
-    status: 'matched' as const,
-    confidence: 1,
-    evidence: ['source-preserved-raster', 'source-text-alt'],
-    candidates: [],
-    sourceBoxes: [captionBox, sourceBox],
-    sourceText,
-    altText: 'Equation shared. Source-backed display.',
-    altTextSource: 'source-text' as const,
-    canonicalNodeId: nodeId,
-    captionNodeId,
-  } satisfies PublicationVisualRelationship
-  return {
-    region,
-    relationship,
-    nodes: [
-      {
-        id: nodeId,
-        type: 'figure' as const,
-        objectType: 'equation' as const,
-        title: relationship.altText,
-        relationships: {
-          caption: captionNodeId,
-          assets: [figure.asset.id],
-        },
-        source: 'synthetic-shared-cross-kind-equation',
-      },
-      {
-        id: captionNodeId,
-        type: 'caption' as const,
-        text: relationship.altText,
-        source: 'synthetic-shared-cross-kind-equation',
-      },
-    ],
-    provenance: {
-      [nodeId]: {
-        confidence: 1,
-        pages: [1],
-        regionIds: [sourceRegionId],
-        boxes: relationship.sourceBoxes,
-        links: [],
-      },
-      [captionNodeId]: {
-        confidence: 1,
-        pages: [1],
-        regionIds: [captionRegionId],
-        boxes: [captionBox],
-        links: [],
-      },
-    },
-  }
 }
 
 function exportBoundaryRun(
@@ -5697,967 +4898,6 @@ describe('EPUB 3 export', () => {
         semanticKind: 'algorithm',
       }),
     ).not.toContain('data-wide-source-visual="true"')
-  })
-
-  it('builds an explicitly non-publication-grade readable fallback without weakening the strict gate', async () => {
-    const run: PdfSourceRun = {
-      page: 1,
-      text: 'Readable text must still flow when a decorative image is unresolved.',
-      x: 0.1,
-      y: 0.2,
-      width: 0.7,
-      height: 0.02,
-      rotation: 0,
-      method: 'pdf-text',
-      fontName: 'Body',
-      fontSize: 10,
-      confidence: 1,
-    }
-    const page: PdfPageAnalysis = {
-      page: 1,
-      kind: 'born-digital',
-      width: 612,
-      height: 792,
-      rotation: 0,
-      textCharacters: run.text.length,
-      imageCount: 1,
-      objects: [],
-      runs: [run],
-    }
-    const reconstruction = await reconstructPageAnalyses({
-      pages: [page],
-      sourceHash: 'a'.repeat(64),
-      fileName: 'readable-fallback.pdf',
-      byteLength: 1024,
-    })
-    reconstruction.paper.title = run.text
-
-    expect(reconstruction.readiness.ready).toBe(false)
-    await expect(
-      buildEpub(reconstruction.paper, reconstruction),
-    ).rejects.toMatchObject({ code: 'INCOMPLETE_RECONSTRUCTION' })
-
-    const fallback = await buildReadableEpub(
-      reconstruction.paper,
-      reconstruction,
-    )
-    const { files, manifest } = inspectEpub(fallback.bytes)
-    const content = strFromU8(files['EPUB/content.xhtml'])
-
-    expect(fallback.mode).toBe('readable-fallback')
-    expect(fallback.fileName).toMatch(/-readable\.epub$/)
-    expect(content).toContain('Readable text must still flow')
-    expect(content).toContain(
-      '<title>Readable text must still flow when a decorative image is unresolved.</title>',
-    )
-    expect(content).not.toContain('<title>Publication</title>')
-    expect(content.match(/Readable text must still flow/g)).toHaveLength(2)
-    expect(content).not.toContain(
-      'This readable fallback is incomplete and is not publication-grade.',
-    )
-    expect(content).not.toContain(
-      'Omitted source visuals and unresolved relationships require review against the source PDF.',
-    )
-    expect(content).not.toContain('class="reconstruction-status"')
-    expect(content).not.toContain('class="publication-header"')
-    expect(content).not.toContain('class="reconstructed-header"')
-    expect(content).not.toContain('class="reconstructed-header"')
-    expect(content).not.toContain('working · 1970-01-01')
-    expect(manifest).toMatchObject({
-      exportMode: 'readable-fallback',
-      publicationGrade: false,
-      sourceReadiness: { ready: false },
-    })
-  })
-
-  it('refuses a readable fallback for a sparse mixed raster page without OCR evidence', async () => {
-    const run: PdfSourceRun = {
-      page: 1,
-      text: 'Sparse embedded heading',
-      x: 0.1,
-      y: 0.1,
-      width: 0.35,
-      height: 0.03,
-      rotation: 0,
-      method: 'pdf-text',
-      fontName: 'Body',
-      fontSize: 12,
-      confidence: 1,
-    }
-    const reconstruction = await reconstructPageAnalyses({
-      pages: [
-        {
-          page: 1,
-          kind: 'mixed',
-          width: 612,
-          height: 792,
-          rotation: 0,
-          textCharacters: run.text.length,
-          imageCount: 1,
-          objects: [],
-          runs: [run],
-        },
-      ],
-      sourceHash: 'b'.repeat(64),
-      fileName: 'mixed-without-ocr.pdf',
-      byteLength: 1024,
-    })
-
-    expect(reconstruction.completeness.ocrRequiredPages).toEqual([1])
-    await expect(
-      buildReadableEpub(reconstruction.paper, reconstruction),
-    ).rejects.toMatchObject({
-      code: 'INCOMPLETE_RECONSTRUCTION',
-    })
-  })
-
-  it('builds a bounded source-preserved page fallback when a scan has no recovered text', async () => {
-    const reconstruction = await reconstructPdf(
-      await fixtureFile('scanned-page.pdf'),
-    )
-
-    expect(reconstruction.completeness.ocrRequiredPages).toEqual([1])
-    const fallback = await buildReadableEpub(
-      reconstruction.paper,
-      reconstruction,
-    )
-    const { files, manifest } = inspectEpub(fallback.bytes)
-    const content = strFromU8(files['EPUB/content.xhtml'])
-
-    expect(fallback.mode).toBe('readable-fallback')
-    expect(content).toContain('Source page 1')
-    expect(content).toContain('text recovery required')
-    expect(content).toContain('<img')
-    expect(manifest).toMatchObject({
-      publicationGrade: false,
-      sourceCompleteness: { ocrRequiredPages: [1] },
-      excludedUnresolvedVisualRelationshipCount: 0,
-    })
-    expect(manifest.assets).toHaveLength(1)
-    const sourceAsset = reconstruction.pages[0]!.assets!.find(
-      (asset) => asset.rendition === 'source-page-render',
-    )!
-    expect(manifest.assets[0]).toMatchObject({ sourceAssetId: sourceAsset.id })
-    expect(files[`EPUB/${sourceAsset.href}`]).toBeInstanceOf(Uint8Array)
-  })
-
-  it.each([
-    {
-      name: 'an oversized unselected ordinary top-level asset',
-      ordinaryAssets: (source: PublicationAsset) => [
-        {
-          ...source,
-          id: 'ordinary-oversized-top-level',
-          bytes: {
-            byteLength: MAX_EPUB_ASSET_BYTES_PER_BOOK + 1,
-          } as unknown as Uint8Array,
-        },
-      ],
-    },
-    {
-      name: '512 unselected ordinary top-level assets',
-      ordinaryAssets: (source: PublicationAsset) =>
-        Array.from({ length: MAX_EPUB_ASSETS_PER_BOOK }, (_, index) => ({
-          ...source,
-          id: `ordinary-top-level-${index + 1}`,
-        })),
-    },
-  ])(
-    'keeps an unresolved page render bounded when reconstruction also has $name',
-    async ({ ordinaryAssets }) => {
-      const reconstruction = await reconstructPdf(
-        await fixtureFile('scanned-page.pdf'),
-      )
-      const sourceRender = reconstruction.pages[0]!.assets!.find(
-        (asset) => asset.rendition === 'source-page-render',
-      )!
-      const sourceObject = reconstruction.pages[0]!.objects!.find(
-        (object) => object.rolePolicy === 'pdfjs-complete-page-render-v1',
-      )!
-      const ordinarySource = reconstruction.pages[0]!.assets!.find(
-        (asset) => asset.rendition === 'source-preserved',
-      )!
-      reconstruction.pages[0]!.assets = reconstruction.pages[0]!.assets!.filter(
-        (asset) => asset.id !== sourceRender.id,
-      )
-      reconstruction.pages[0]!.objects =
-        reconstruction.pages[0]!.objects!.filter(
-          (object) => object.id !== sourceObject.id,
-        )
-      reconstruction.assets = ordinaryAssets(ordinarySource)
-
-      expect(reconstruction.completeness.ocrRequiredPages).toEqual([1])
-      expect(
-        reconstruction.pages[0]!.assets!.some(
-          (asset) => asset.rendition === 'source-page-render',
-        ),
-      ).toBe(false)
-      await attachRequiredSourcePageRenders(
-        reconstruction,
-        async (target, maximumBytes) => {
-          expect(maximumBytes).toBeGreaterThanOrEqual(
-            sourceRender.bytes.byteLength,
-          )
-          target.assets = [...(target.assets ?? []), sourceRender]
-          target.objects = [...(target.objects ?? []), sourceObject]
-          return sourceRender.bytes.byteLength
-        },
-      )
-      expect(
-        reconstruction.pages[0]!.assets!.some(
-          (asset) => asset.id === sourceRender.id,
-        ),
-      ).toBe(true)
-      expect(
-        sourcePageRenderBudgetUsage(
-          reconstruction.assets,
-          reconstruction.pages,
-        ),
-      ).toEqual({
-        count: 1,
-        byteLength: sourceRender.bytes.byteLength,
-      })
-
-      const projected = projectReadableFallbackReconstruction(reconstruction)
-      expect(projected.assets).toEqual([sourceRender])
-      expect(projected.assets.length).toBeLessThanOrEqual(
-        MAX_EPUB_ASSETS_PER_BOOK,
-      )
-      expect(
-        projected.assets.reduce(
-          (total, asset) => total + asset.bytes.byteLength,
-          0,
-        ),
-      ).toBeLessThanOrEqual(MAX_EPUB_ASSET_BYTES_PER_BOOK)
-    },
-    120_000,
-  )
-
-  it.each([
-    'scanned-page.pdf',
-    'rotated-scan.pdf',
-    'two-page-scan.pdf',
-    'two-physical-page-scan.pdf',
-    'tiled-multi-image-scan.pdf',
-    'multilingual-scan.pdf',
-  ])(
-    'uses complete rendered source pages with closed fallback lineage for %s',
-    async (name) => {
-      const reconstruction = await reconstructPdf(await fixtureFile(name))
-      const projected = projectReadableFallbackReconstruction(reconstruction)
-
-      expect(projected.assets).toHaveLength(reconstruction.pages.length)
-      for (const page of reconstruction.pages) {
-        const surface = page.objects?.find(
-          (object) =>
-            object.role === 'scan-source' &&
-            object.rolePolicy === 'pdfjs-complete-page-render-v1' &&
-            object.box.x === 0 &&
-            object.box.y === 0 &&
-            object.box.width === 1 &&
-            object.box.height === 1,
-        )
-        expect(surface).toBeDefined()
-        const sourceAsset = page.assets?.find(
-          (asset) => asset.id === surface?.assetId,
-        )
-        expect(sourceAsset).toMatchObject({
-          rendition: 'source-page-render',
-          mediaType: 'image/png',
-          sourceObjectIds: [surface!.id],
-          sourceBoxes: [surface!.box],
-        })
-        expect(sourceAsset!.width / sourceAsset!.height).toBeCloseTo(
-          page.width / page.height,
-          2,
-        )
-        expect(sourceAsset!.sourceBoxes[0]!.rotation).toBe(page.rotation)
-        expect(
-          projected.assets.some((asset) => asset.id === sourceAsset!.id),
-        ).toBe(true)
-      }
-      for (const relationship of projected.visualRelationships) {
-        expect(
-          projected.regions.some(
-            (region) => region.id === relationship.captionRegionId,
-          ),
-        ).toBe(true)
-        expect(projected.provenance[relationship.captionNodeId!]).toMatchObject(
-          {
-            confidence: 0,
-            regionIds: [relationship.captionRegionId],
-          },
-        )
-      }
-    },
-    120_000,
-  )
-
-  it.each(['mixed-page.pdf', 'sparse-embedded-text.pdf'])(
-    'preserves readable text and appends a complete-page figure for unresolved %s',
-    async (name) => {
-      const reconstruction = await reconstructPdf(await fixtureFile(name))
-      const readableNodeIds = reconstruction.paper.nodes
-        .filter((node) => 'text' in node && node.text.trim())
-        .map((node) => node.id)
-      const projected = projectReadableFallbackReconstruction(reconstruction)
-
-      expect(reconstruction.completeness.ocrRequiredPages).toEqual([1])
-      expect(readableNodeIds.length).toBeGreaterThan(0)
-      expect(projected.paper.nodes.map((node) => node.id)).toEqual(
-        expect.arrayContaining(readableNodeIds),
-      )
-      expect(
-        projected.paper.nodes.filter(
-          (node) =>
-            node.type === 'figure' && node.id === 'source-scan-page-001',
-        ),
-      ).toHaveLength(1)
-      expect(
-        projected.assets.filter(
-          (asset) => asset.rendition === 'source-page-render',
-        ),
-      ).toHaveLength(1)
-      const fallback = await buildReadableEpub(
-        reconstruction.paper,
-        reconstruction,
-      )
-      const { files } = inspectEpub(fallback.bytes)
-      const content = strFromU8(files['EPUB/content.xhtml'])
-      expect(content).toContain('Source page 1')
-      expect(content).toContain('<img')
-    },
-    120_000,
-  )
-
-  it('keeps a resolved digital page readable and renders only the unresolved physical page', async () => {
-    const reconstruction = await reconstructPdf(
-      await fixtureFile('mixed-digital-scan.pdf'),
-    )
-    const projected = projectReadableFallbackReconstruction(reconstruction)
-
-    expect(reconstruction.completeness.ocrRequiredPages).toEqual([2])
-    expect(
-      reconstruction.pages[0]!.assets?.some(
-        (asset) => asset.rendition === 'source-page-render',
-      ) ?? false,
-    ).toBe(false)
-    expect(
-      reconstruction.pages[1]!.assets?.some(
-        (asset) => asset.rendition === 'source-page-render',
-      ),
-    ).toBe(true)
-    expect(
-      projected.paper.nodes.some(
-        (node) =>
-          'text' in node && node.text.includes('Readable digital introduction'),
-      ),
-    ).toBe(true)
-    expect(
-      projected.paper.nodes
-        .filter(
-          (node) =>
-            node.type === 'figure' && node.id.startsWith('source-scan-page-'),
-        )
-        .map((node) => node.id),
-    ).toEqual(['source-scan-page-002'])
-  }, 120_000)
-
-  it('places a scan-first fallback before readable nodes from later physical pages', async () => {
-    const reconstruction = await reconstructPdf(
-      await fixtureFile('scan-digital-hybrid.pdf'),
-    )
-    const projected = projectReadableFallbackReconstruction(reconstruction)
-    const nodeIds = projected.paper.nodes.map((node) => node.id)
-    const fallbackIndex = nodeIds.indexOf('source-scan-page-001')
-    const laterTextIndex = projected.paper.nodes.findIndex(
-      (node) =>
-        'text' in node && node.text.includes('Readable digital conclusion'),
-    )
-
-    expect(reconstruction.completeness.ocrRequiredPages).toEqual([1])
-    expect(fallbackIndex).toBeGreaterThanOrEqual(0)
-    expect(laterTextIndex).toBeGreaterThan(fallbackIndex)
-  }, 120_000)
-
-  it('uses a complete page render instead of a large embedded image when the page adds an overlay', async () => {
-    const source = await fixtureFile('scanned-page.pdf')
-    const sourceDocument = await PDFDocument.load(await source.arrayBuffer())
-    const overlaidDocument = await PDFDocument.create()
-    const [page] = await overlaidDocument.copyPages(sourceDocument, [0])
-    overlaidDocument.addPage(page)
-    page.drawRectangle({
-      x: 12,
-      y: 12,
-      width: 72,
-      height: 36,
-      borderColor: rgb(0.8, 0.1, 0.1),
-      borderWidth: 5,
-    })
-    const bytes = await overlaidDocument.save({ useObjectStreams: false })
-    const reconstruction = await reconstructPdf(
-      new File([bytes], 'unseen-overlay-scan.pdf', {
-        type: 'application/pdf',
-        lastModified: 0,
-      }),
-    )
-    const projected = projectReadableFallbackReconstruction(reconstruction)
-    const nativeAssets = reconstruction.pages[0]!.assets!.filter(
-      (asset) => asset.rendition === 'source-preserved',
-    )
-
-    expect(nativeAssets.length).toBeGreaterThan(0)
-    expect(projected.assets).toHaveLength(1)
-    expect(projected.assets[0]).toMatchObject({
-      rendition: 'source-page-render',
-    })
-    expect(
-      nativeAssets.some(
-        (asset) => asset.sha256 === projected.assets[0]!.sha256,
-      ),
-    ).toBe(false)
-  })
-
-  it.each([
-    {
-      name: 'a derived rendition',
-      mutate: (reconstruction: PdfReconstruction) => {
-        reconstruction.pages[0]!.assets!.find(
-          (asset) => asset.rendition === 'source-page-render',
-        )!.rendition = 'profile-downscaled'
-      },
-    },
-    {
-      name: 'content that no longer matches its digest',
-      mutate: (reconstruction: PdfReconstruction) => {
-        reconstruction.pages[0]!.assets!.find(
-          (asset) => asset.rendition === 'source-page-render',
-        )!.bytes[0] ^= 0xff
-      },
-    },
-    {
-      name: 'an href that no longer matches its digest',
-      mutate: (reconstruction: PdfReconstruction) => {
-        reconstruction.pages[0]!.assets!.find(
-          (asset) => asset.rendition === 'source-page-render',
-        )!.href = 'assets/forged-source-page.png'
-      },
-    },
-  ])('refuses a scan fallback backed by $name', async ({ mutate }) => {
-    const reconstruction = await reconstructPdf(
-      await fixtureFile('scanned-page.pdf'),
-    )
-    mutate(reconstruction)
-
-    await expect(
-      buildReadableEpub(reconstruction.paper, reconstruction),
-    ).rejects.toMatchObject({ code: 'INCOMPLETE_RECONSTRUCTION' })
-  })
-
-  it('keeps a complete bounded source-backed table image in readable fallback', async () => {
-    const run: PdfSourceRun = {
-      page: 1,
-      text: 'Readable prose around a source-backed table.',
-      x: 0.1,
-      y: 0.1,
-      width: 0.7,
-      height: 0.02,
-      rotation: 0,
-      method: 'pdf-text',
-      fontName: 'Body',
-      fontSize: 10,
-      confidence: 1,
-    }
-    const reconstruction = await reconstructPageAnalyses({
-      pages: [
-        {
-          page: 1,
-          kind: 'born-digital',
-          width: 612,
-          height: 792,
-          rotation: 0,
-          textCharacters: run.text.length,
-          imageCount: 1,
-          objects: [],
-          runs: [run],
-        },
-      ],
-      sourceHash: '7'.repeat(64),
-      fileName: 'bounded-table-fallback.pdf',
-      byteLength: 1024,
-    })
-    const captionBox = {
-      page: 1,
-      x: 0.15,
-      y: 0.28,
-      width: 0.7,
-      height: 0.03,
-      rotation: 0,
-      method: 'pdf-text' as const,
-    }
-    const tableBox = {
-      page: 1,
-      x: 0.18,
-      y: 0.33,
-      width: 0.64,
-      height: 0.24,
-      rotation: 0,
-      method: 'pdf-text' as const,
-    }
-    const sourceTableWidth = 1_000
-    const sourceTableHeight = 900
-    const crop = await createSourcePageCropAsset({
-      kind: 'table',
-      cropBox: { ...tableBox, method: 'pdf-object' },
-      sourceObjectIds: ['bounded-table-source'],
-      sourceBoxes: [tableBox],
-      width: sourceTableWidth,
-      height: sourceTableHeight,
-      pixels: new Uint8Array(sourceTableWidth * sourceTableHeight * 4).fill(80),
-    })
-    const sourceText = 'Method Score Baseline 72 Proposed 81'
-    const figureNode = {
-      id: 'bounded-table-node',
-      type: 'figure' as const,
-      objectType: 'table' as const,
-      title: 'Table 1. Source-backed comparison.',
-      sourceText,
-      inlineRuns: [
-        { start: 0, end: 'Method'.length, bold: true },
-        {
-          start: sourceText.indexOf('Proposed'),
-          end: sourceText.indexOf('Proposed') + 'Proposed'.length,
-          italic: true,
-        },
-      ],
-      relationships: {
-        caption: 'bounded-table-caption',
-        assets: [crop.id],
-      },
-      source: 'synthetic-bounded-table',
-    }
-    const captionNode = {
-      id: 'bounded-table-caption',
-      type: 'caption' as const,
-      text: 'Table 1. Source-backed comparison.',
-      source: 'synthetic-bounded-table',
-    }
-    const relationship = {
-      id: 'bounded-table-relationship',
-      kind: 'table' as const,
-      label: 'Table 1',
-      captionRegionId: 'bounded-table-caption-region',
-      sourceRegionIds: ['bounded-table-source-region'],
-      sourceLineIds: ['bounded-table-header-line', 'bounded-table-body-line'],
-      sourceObjectIds: ['bounded-table-source'],
-      assetIds: [crop.id],
-      status: 'matched' as const,
-      confidence: 1,
-      evidence: [
-        'bounded-table-scope',
-        'non-semantic-source-scope',
-        'source-page-crop',
-      ],
-      candidates: [],
-      sourceBoxes: [captionBox, tableBox],
-      sourceText,
-      altText: captionNode.text,
-      altTextSource: 'caption' as const,
-      canonicalNodeId: figureNode.id,
-      captionNodeId: captionNode.id,
-    }
-    const withTable = {
-      ...reconstruction,
-      paper: {
-        ...reconstruction.paper,
-        nodes: [...reconstruction.paper.nodes, figureNode, captionNode],
-      },
-      provenance: {
-        ...reconstruction.provenance,
-        [figureNode.id]: {
-          confidence: 1,
-          pages: [1],
-          regionIds: relationship.sourceRegionIds,
-          boxes: relationship.sourceBoxes,
-          links: [],
-        },
-        [captionNode.id]: {
-          confidence: 1,
-          pages: [1],
-          regionIds: [relationship.captionRegionId],
-          boxes: [captionBox],
-          links: [],
-        },
-      },
-      assets: [crop],
-      visualRelationships: [relationship],
-    } satisfies PdfReconstruction
-
-    const fallback = await buildReadableEpub(withTable.paper, withTable)
-    const { files, manifest } = inspectEpub(fallback.bytes)
-    const content = strFromU8(files['EPUB/content.xhtml'])
-
-    expect(content).toContain('data-object-type="table"')
-    expect(content).toContain(`data-asset-id="${crop.id}"`)
-    expect(content).not.toContain('loading="lazy"')
-    expect(content).toContain('loading="eager"')
-    expect(content).toContain('Table 1. Source-backed comparison.')
-    expect(content).toContain(
-      '<span class="visually-hidden visual-source-transcript"',
-    )
-    expect(content).toContain(
-      '<strong>Method</strong> Score Baseline 72 <em>Proposed</em> 81',
-    )
-    expect(content).not.toContain('<table')
-    expect(manifest.visualRelationships).toEqual([
-      expect.objectContaining({
-        id: relationship.id,
-        kind: 'table',
-        assetIds: [crop.id],
-      }),
-    ])
-    expect(manifest.assets).toEqual([
-      expect.objectContaining({
-        id: crop.id,
-        kind: 'table',
-        rendition: 'source-page-crop',
-      }),
-    ])
-
-    const moveProfile = getTargetProfile('paperProMove')
-    const moveFallback = await buildReadableEpub(
-      withTable.paper,
-      withTable,
-      moveProfile,
-    )
-    const moveInspection = inspectEpub(moveFallback.bytes, moveProfile)
-    const moveContent = strFromU8(moveInspection.files['EPUB/content.xhtml'])
-    const moveManifest = moveInspection.manifest as {
-      assets?: Array<Record<string, unknown>>
-    }
-    const moveTableAsset = moveManifest.assets?.find(
-      (candidate) => candidate.sourceAssetId === crop.id,
-    )
-
-    expect(moveContent).toContain(
-      'data-wide-source-visual="true" data-source-visual-kind="table"',
-    )
-    expect(moveContent).toContain(
-      `<img src="${crop.href}" width="${sourceTableWidth}" height="${sourceTableHeight}"`,
-    )
-    expect(moveInspection.files[`EPUB/${crop.href}`]).toEqual(crop.bytes)
-    expect(moveTableAsset).toMatchObject({
-      id: crop.id,
-      width: sourceTableWidth,
-      height: sourceTableHeight,
-      sourceAssetId: crop.id,
-      policy: {
-        id: 'preserve-scrollable-table-source',
-        action: 'preserved',
-        sourceWidth: sourceTableWidth,
-        sourceHeight: sourceTableHeight,
-        packagedWidth: sourceTableWidth,
-        packagedHeight: sourceTableHeight,
-        maximumWidth: null,
-        resampling: 'none',
-        neverUpscaled: true,
-      },
-    })
-  })
-
-  it('retains more than 64 unique matched equation assets outside the optional cap', async () => {
-    const { reconstruction } = await staleEquationTranscriptFixture()
-    const equations = await Promise.all(
-      Array.from({ length: 65 }, (_, index) =>
-        readableFallbackEquation(index + 1),
-      ),
-    )
-    const projected = projectReadableFallbackReconstruction(
-      withReadableFallbackEquations(reconstruction, equations),
-    )
-    const equationRelationships = projected.visualRelationships.filter(
-      (relationship) => relationship.kind === 'equation',
-    )
-    const equationAssetIds = new Set(
-      equationRelationships.flatMap((relationship) => relationship.assetIds),
-    )
-
-    expect(equationRelationships).toHaveLength(66)
-    expect(equationAssetIds).toHaveLength(66)
-    expect(projected.assets).toHaveLength(66)
-    const epub = await buildReadableEpub(
-      projected.paper,
-      projected,
-      getTargetProfile('paperPro'),
-    )
-    const content = strFromU8(
-      inspectEpub(epub.bytes, getTargetProfile('paperPro')).files[
-        'EPUB/content.xhtml'
-      ],
-    )
-    expect(content.match(/data-object-type="equation"/gu)).toHaveLength(66)
-    for (const relationship of equationRelationships) {
-      expect(content).toContain(
-        `id="${relationship.canonicalNodeId}" data-canonical-id="${relationship.canonicalNodeId}"`,
-      )
-    }
-  })
-
-  it('fails closed when relationships share a canonical or caption owner', async () => {
-    const { reconstruction, relationship } =
-      await staleEquationTranscriptFixture()
-    for (const duplicate of [
-      {
-        ...relationship,
-        id: 'duplicate-equation-canonical-owner',
-        captionNodeId: 'independent-caption-owner',
-      },
-      {
-        ...relationship,
-        id: 'duplicate-equation-caption-owner',
-        canonicalNodeId: 'independent-canonical-owner',
-      },
-    ]) {
-      expect(() =>
-        projectReadableFallbackReconstruction({
-          ...reconstruction,
-          visualRelationships: [relationship, duplicate],
-        }),
-      ).toThrow(/ambiguously share/u)
-    }
-  })
-
-  it('fails before packaging an unbounded mandatory equation workset', async () => {
-    const { reconstruction, relationship } =
-      await staleEquationTranscriptFixture()
-    const excessiveRelationships = Array.from(
-      { length: MAX_READABLE_FALLBACK_EQUATIONS_PER_BOOK + 1 },
-      (_, index) => ({
-        ...relationship,
-        id: `resource-bounded-equation-${index + 1}`,
-      }),
-    )
-    expect(() =>
-      projectReadableFallbackReconstruction({
-        ...reconstruction,
-        visualRelationships: excessiveRelationships,
-      }),
-    ).toThrow(/EPUB asset resource limit exceeded/u)
-
-    expect(() =>
-      projectReadableFallbackReconstruction({
-        ...reconstruction,
-        assets: [
-          {
-            ...reconstruction.assets[0],
-            bytes: {
-              byteLength: MAX_EPUB_ASSET_BYTES_PER_BOOK + 1,
-            } as unknown as Uint8Array,
-          },
-        ],
-      }),
-    ).toThrow(/EPUB asset resource limit exceeded/u)
-  })
-
-  it('caps unique optional asset IDs without collapsing relationship ownership', async () => {
-    const { reconstruction } = await staleEquationTranscriptFixture()
-    const uniqueFigures = Array.from({ length: 65 }, (_, index) =>
-      readableFallbackFigure(index + 1),
-    )
-    const uniqueProjection = projectReadableFallbackReconstruction(
-      withReadableFallbackFigures(reconstruction, uniqueFigures),
-    )
-
-    expect(
-      uniqueProjection.visualRelationships.filter(
-        (relationship) => relationship.kind === 'equation',
-      ),
-    ).toHaveLength(1)
-    expect(
-      uniqueProjection.visualRelationships.filter(
-        (relationship) => relationship.kind === 'figure',
-      ),
-    ).toHaveLength(64)
-
-    const crossKindSharedFigure = readableFallbackFigure(1)
-    const crossKindOptionalFigures = [
-      crossKindSharedFigure,
-      ...Array.from({ length: 64 }, (_, index) =>
-        readableFallbackFigure(index + 2),
-      ),
-    ]
-    const crossKindEquation = readableFallbackEquationSharingFigureAsset(
-      crossKindSharedFigure,
-    )
-    const crossKindBase = withReadableFallbackFigures(
-      reconstruction,
-      crossKindOptionalFigures,
-    )
-    const crossKindProjection = projectReadableFallbackReconstruction({
-      ...crossKindBase,
-      paper: {
-        ...crossKindBase.paper,
-        nodes: [...crossKindBase.paper.nodes, ...crossKindEquation.nodes],
-      },
-      provenance: {
-        ...crossKindBase.provenance,
-        ...crossKindEquation.provenance,
-      },
-      regions: [...crossKindBase.regions, crossKindEquation.region],
-      visualRelationships: [
-        ...reconstruction.visualRelationships,
-        ...crossKindOptionalFigures.map((figure) => figure.relationship),
-        crossKindEquation.relationship,
-      ],
-    })
-
-    expect(
-      crossKindProjection.visualRelationships.filter(
-        (relationship) => relationship.kind === 'figure',
-      ),
-    ).toHaveLength(65)
-    expect(crossKindProjection.assets).toHaveLength(66)
-  })
-
-  it('fails readable fallback explicitly for missing or invalid matched equation assets', async () => {
-    const { reconstruction, relationship } =
-      await staleEquationTranscriptFixture()
-    const invalidReconstructions = [
-      {
-        name: 'missing',
-        reconstruction: { ...reconstruction, assets: [] },
-      },
-      {
-        name: 'invalid',
-        reconstruction: {
-          ...reconstruction,
-          assets: [
-            {
-              ...reconstruction.assets[0],
-              sha256: '0'.repeat(64),
-            },
-          ],
-        },
-      },
-    ] satisfies Array<{ name: string; reconstruction: PdfReconstruction }>
-
-    for (const fixture of invalidReconstructions) {
-      let error: unknown
-      try {
-        projectReadableFallbackReconstruction(fixture.reconstruction)
-      } catch (caught) {
-        error = caught
-      }
-      expect(error, fixture.name).toMatchObject({
-        code: 'INCOMPLETE_RECONSTRUCTION',
-      })
-      expect(String(error)).toContain(relationship.id)
-    }
-  })
-
-  it('projects readable-fallback note backlinks onto anchors that can actually render', async () => {
-    const run: PdfSourceRun = {
-      page: 1,
-      text: 'A claim with note 1.',
-      x: 0.1,
-      y: 0.2,
-      width: 0.7,
-      height: 0.02,
-      rotation: 0,
-      method: 'pdf-text',
-      fontName: 'Body',
-      fontSize: 10,
-      confidence: 1,
-    }
-    const reconstruction = await reconstructPageAnalyses({
-      pages: [
-        {
-          page: 1,
-          kind: 'born-digital',
-          width: 612,
-          height: 792,
-          rotation: 0,
-          textCharacters: run.text.length,
-          imageCount: 1,
-          objects: [],
-          runs: [run],
-        },
-      ],
-      sourceHash: 'b'.repeat(64),
-      fileName: 'readable-note-projection.pdf',
-      byteLength: 1024,
-    })
-    const paragraph = reconstruction.paper.nodes.find(
-      (node) => node.type === 'paragraph' && node.text === run.text,
-    )
-    expect(paragraph?.type).toBe('paragraph')
-    if (paragraph?.type !== 'paragraph') throw new Error('missing paragraph')
-    paragraph.noteReferences = [
-      {
-        id: 'rendered-reference',
-        label: '1',
-        target: 'rendered-note',
-        start: run.text.indexOf('1'),
-        end: run.text.indexOf('1') + 1,
-        confidence: 1,
-      },
-    ]
-    reconstruction.paper.nodes.push(
-      {
-        id: 'rendered-note',
-        type: 'footnote',
-        kind: 'footnote',
-        label: '1',
-        text: 'A source-backed note.',
-        relationships: {
-          backlinks: ['rendered-reference', 'orphan-source-marker'],
-        },
-        source: 'pdf:synthetic#page=1',
-      },
-      {
-        id: 'orphan-note',
-        type: 'footnote',
-        kind: 'footnote',
-        label: '2',
-        text: 'A note whose extracted marker cannot render.',
-        relationships: { backlinks: ['another-orphan-source-marker'] },
-        source: 'pdf:synthetic#page=1',
-      },
-    )
-    const referenceStart = run.text.indexOf('1')
-    reconstruction.noteRelationships.push({
-      id: 'rendered-reference',
-      label: '1',
-      referenceRegionId: reconstruction.provenance[paragraph.id].regionIds[0],
-      referenceStart,
-      referenceEnd: referenceStart + 1,
-      targetNoteId: 'rendered-note',
-      status: 'matched',
-      canonicalAnchor: {
-        kind: 'node',
-        nodeId: paragraph.id,
-        start: referenceStart,
-        end: referenceStart + 1,
-      },
-      confidence: 1,
-      threshold: 0.72,
-      evidence: ['synthetic-rendered-note-reference'],
-      candidates: [],
-      sourceBoxes: reconstruction.provenance[paragraph.id].boxes,
-    })
-
-    const fallback = await buildReadableEpub(
-      reconstruction.paper,
-      reconstruction,
-    )
-    const { files } = inspectEpub(fallback.bytes)
-    const content = strFromU8(files['EPUB/content.xhtml'])
-
-    expect(content).toContain('id="rendered-reference"')
-    expect(content).toContain('href="#rendered-reference"')
-    expect(content).not.toContain('orphan-source-marker')
-    expect(content).not.toContain('another-orphan-source-marker')
-    expect(content).toContain('A note whose extracted marker cannot render.')
   })
 
   it('keeps unresolved diagram OCR out of reader-facing content', () => {
